Revert scaling and add an offset variable to allow VC31 HRM values to fit into a more sensible range, allow HRM to work when used immediately after a load()

+++ libs/misc/heartrate.h
@@ -20,7 +20,7 @@
 #define HRM_THRESH_SHIFT 4
 
 typedef struct {
-  int16_t raw;
+  int8_t raw;
   int8_t filtered;
   int8_t thresh;
   bool wasLow;
